clean up cache operations and generalize

+++ src/steps/cache.py
@@ -1,7 +1,11 @@
+from datetime import datetime, timedelta
 import os
 import hashlib
 import shutil
 import botocore
+
+# Cache expiration time
+NEXT_MONTH = datetime.now() + timedelta(days=30)
 
 
 def get_checksum(filename):
@@ -17,10 +21,13 @@ class CacheFolder():
     An abstract class for a cache folder in S3
     '''
 
-    def __init__(self, key, bucket, s3_client):
-        self.key = key
+    def __init__(self, checksum_file, local_folder, bucket, s3_client, logger):
+        self.checksum_file = checksum_file
+        self.key = get_checksum(checksum_file)
+        self.local_folder = local_folder
         self.bucket = bucket
         self.s3_client = s3_client
+        self.logger = logger
 
     def exists(self):
         '''Check if a given cache key exists'''
@@ -34,25 +41,31 @@ class CacheFolder():
             if error.response['Error']['Message'] == 'Not Found':
                 return False
             else:
-                print(error.response['Error'])
+                self.logger.error(error.response['Error'])
                 raise error
 
-    def zip_upload_folder_to_s3(self, folder):
+    def zip_upload_folder_to_s3(self):
+        self.logger.info(f'Caching dependencies from {self.local_folder}.')
         tmp_file = f'{self.key}.zip'
-        shutil.make_archive(self.key, 'zip', folder)
+        shutil.make_archive(self.key, 'zip', self.local_folder)
         self.s3_client.upload_file(
             Filename=tmp_file,
             Bucket=self.bucket,
             Key=f'_cache/{self.key}',
+            ExtraArgs=dict(Expires=NEXT_MONTH)
         )
         os.unlink(tmp_file)
 
-    def download_unzip(self, folder):
-        tmp_file = f'{self.key}.zip'
-        self.s3_client.download_file(
-            Filename=tmp_file,
-            Bucket=self.bucket,
-            Key=f'_cache/{self.key}'
-        )
-        shutil.unpack_archive(tmp_file, folder, 'zip')
-        os.unlink(tmp_file)
+    def download_unzip(self):
+        if self.exists():
+            self.logger.info(f'Dependency cache found, downloading to {self.local_folder}.')
+            tmp_file = f'{self.key}.zip'
+            self.s3_client.download_file(
+                Filename=tmp_file,
+                Bucket=self.bucket,
+                Key=f'_cache/{self.key}'
+            )
+            shutil.unpack_archive(tmp_file, self.local_folder, 'zip')
+            os.unlink(tmp_file)
+        else:
+            self.logger.info('No cache file found.')
